+++ wbooks/src/app/screens/Library/screens/BookList/styles.ts
@@ -1,0 +1,13 @@
+import { StyleSheet, ViewStyle } from 'react-native';
+import { COLORS } from '@constants/colors';
+
+interface Styles {
+  container: ViewStyle;
+}
+
+export default StyleSheet.create<Styles>({
+  container: {
+    backgroundColor: COLORS.lightBlue,
+    padding: 15
+  }
+});
