+++ wbooks/src/constants/fonts.ts
@@ -1,0 +1,6 @@
+export const FONTS = {
+  size: {
+    normal: 16,
+    large: 20
+  }
+};
